% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_border_groups.R
\name{add_v_border_groups}
\alias{add_v_border_groups}
\title{Create vertical border group variable}
\usage{
add_v_border_groups(sheet, formats)
}
\arguments{
\item{sheet}{a tidyxl data frame.}

\item{formats}{the formats of the sheet's workbook.}
}
\description{
An internal fucntion that adds a variable delineating cell groups based on borders of header cells.
}
