% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0000_helper-functions.R
\name{add_variable_if_missing}
\alias{add_variable_if_missing}
\title{get tidyABS components}
\usage{
add_variable_if_missing(sheet, var)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
