% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor.R
\name{anchor}
\alias{anchor}
\title{Filter for a cell from which to select others}
\usage{
anchor(cells, row, col)
}
\arguments{
\item{cells}{Data frame, the cells among which the anchor will be sought}

\item{row}{Numeric, the row of the intended cell}

\item{col}{Numeric, the column of the intended cell}
}
\description{
\code{anchor} protects you from selecting an initial cell that
does not exist.  This is useful when choosing an inital cell from which to
search for others, using \code{\link{offset}} and \code{\link{extend}}.
If the chosen cell does not exist, then cell is generated at the given
row and column, with NA in every column other than 'row' and 'col'.
}
\examples{
cells <- tidytable(purpose$`NNW WNW`)
\dontrun{
  anchor(cells, 0, 1)
}
anchor(cells, 1, 1)
anchor(cells, 100, 1)
}

