% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{append_fmt}
\alias{append_fmt}
\title{Add a format variable}
\usage{
append_fmt(cells, fmt_function)
}
\arguments{
\item{cells}{a data frame created by tidyxl::xlsx_cells}

\item{fmt_function}{fmt_* function.}
}
\description{
Add a format variable.
}
