% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{append_fmt_single}
\alias{append_fmt_single}
\title{Add a format variable}
\usage{
append_fmt_single(cells, fmt_function)
}
\arguments{
\item{cells}{a data frame created by tidyxl::xlsx_cells}

\item{fmt_function}{fmt_* function.}
}
\description{
Add a variable indicating formatting of cells.
}
\examples{

 
library(tidyverse)

# Read in a formatted tidyxl data frame. 

xl_df <- 
 unpivotr_example("worked-examples.xlsx") \%>\% 
 xlsx_cells_fmt(sheets = "pivot-hierarchy")
 
# Add a column indicate the leveling of indenting for each cell and locate data cell.
xl_df <- 
 xl_df \%>\% 
  append_fmt(fmt_alignment_indent) \%>\%
  locate_data(data_type == "numeric") 
# Add annotations for header cells. First for header cells to the left of the table with no indenting, and then for cells for one level of indenting.
xl_df <- 
 xl_df \%>\% 
  locate_if(fmt_alignment_indent == 0, direction = "WNW", name = subject_type) \%>\% 
  locate_if(fmt_alignment_indent == 1, direction = "W", name = subject) \%>\% 
  locate(direction = "N", name = student) 
  
# Use `migrate` to reshape the data frame such that each data cells has its own row and each header variable has its own column.  
 xl_df \%>\% migrate()
}
