% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{append_name_to_quosure}
\alias{append_name_to_quosure}
\title{get tidyABS components}
\usage{
append_name_to_quosure(x, prefix = "grp_")
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
