% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{arrange}
\alias{arrange}
\title{Arrange rows by variables}
\usage{
arrange(.data, ...)
}
\arguments{
\item{...}{Comma separated list of unquoted variable names, or expressions}
}
\description{
Order tbl rows by an expression involving its variables.
}
