% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl_fmts.R
\name{bg_color}
\alias{bg_color}
\title{Add formatting information from the bg_color format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' bg_color formatting.}
\usage{
bg_color(format_id_vec = local_format_id, sheet_format = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_format}{formats}
}
\description{
Add formatting information from the bg_color format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' bg_color formatting.
}
