% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\docType{data}
\name{cols_index}
\alias{cols_index}
\title{An ordered set of excel column indexes.}
\format{A character vector of length 702}
\usage{
cols_index
}
\description{
A character vector containing and ordered set of excel column indexes
}
\keyword{datasets}
