% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_border_groups.R
\name{create_v_border_groups}
\alias{create_v_border_groups}
\title{Create vertical border groups}
\usage{
create_v_border_groups(df, formats)
}
\arguments{
\item{df}{a tidyxl data frame.}

\item{formats}{the formats of the sheet's workbook.}
}
\description{
An internal fucntion that identifies groups based on vertical borders of header cells.
}
