% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\docType{data}
\name{direction_plot_interactive}
\alias{direction_plot_interactive}
\title{Matches directions to unicode arrows for an interative chart}
\format{a dataframe}
\usage{
direction_plot_interactive
}
\description{
A dataframe .direction, .arrow and .rate columns
}
\keyword{datasets}
