% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{enhead_tabledata}
\alias{enhead_tabledata}
\title{This is an internal function to wrap \code{enhead}. It takes header cells, data cells and a direction and combines them into a tidy data frame.}
\usage{
enhead_tabledata(header_data, direction, values = tabledata)
}
\arguments{
\item{header_data}{a data frame containing header cells.}

\item{direction}{direction suplied to enhead.}

\item{values}{the data values.}
}
\description{
This is an internal function to wrap \code{enhead}. It takes header cells, data cells and a direction and combines them into a tidy data frame.
}
