% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{fill_blanks_in_headers}
\alias{fill_blanks_in_headers}
\title{Fill in blanks}
\usage{
fill_blanks_in_headers(header_df, header_fill, formats, direction)
}
\arguments{
\item{df}{a data frame containing header cells.}

\item{form_list}{list of fomulae}

\item{format}{the formats associated with the workbook containing the header_df cells.}
}
\description{
This function ensures that merged cells are unmerged.
}
