% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0000_fill_blanks_in_headers.R
\name{fill_blanks_in_row_headers}
\alias{fill_blanks_in_row_headers}
\title{Fill in blanks}
\usage{
fill_blanks_in_row_headers(header_df, header_fill, formats)
}
\arguments{
\item{header_df}{a data frame containing header cells.}

\item{formats}{the formats associated with the workbook containing the header_df cells.}

\item{col_header_fill}{the manner is which blank cells will be filled.}
}
\description{
This function ensures that merged cells are unmerged.
}
