% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_blanks.R
\name{fill_in_blanks}
\alias{fill_in_blanks}
\title{Fill in blanks}
\usage{
fill_in_blanks(sheet)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}
}
\description{
This function ensures that merged cells are unmerged.
}
