% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{filter}
\alias{filter}
\title{Return rows with matching conditions}
\usage{
filter(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{A tbl. All main verbs are S3 generics and provide methods
for \code{\link[=tbl_df]{tbl_df()}}, \code{\link[dtplyr:tbl_dt]{dtplyr::tbl_dt()}} and \code{\link[dbplyr:tbl_dbi]{dbplyr::tbl_dbi()}}.}

\item{...}{Logical predicates defined in terms of the variables in \code{.data}.}
}
\description{
Use \code{filter()} to choose rows/cases where conditions are true. Unlike
base subsetting with \code{[}, rows where the condition evaluates to \code{NA} are
}
