% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{filter.tidyxl}
\alias{filter.tidyxl}
\title{Filter keeping formats}
\usage{
\method{filter}{tidyxl}(df, ...)
}
\arguments{
\item{df}{a data frame created by tidyxl::xlsx_cells}

\item{...}{filter expression.}
}
\description{
A wrapper for dplyr::filter that retains formatting information
}
