% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{filter_fmt}
\alias{filter_fmt}
\title{Filter keeping formats}
\usage{
filter_fmt(df, ...)
}
\arguments{
\item{df}{a data frame created by tidyxl::xlsx_cells}

\item{...}{filter expression.}
}
\description{
A wrapper for dplyr::filter that retains formatting information
}
\examples{
print("todo")
}
