% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmt_vecs.R
\name{fmt_border_top_color_rgb}
\alias{fmt_border_top_color_rgb}
\title{Format function68
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.}
\usage{
fmt_border_top_color_rgb(format_id_vec = local_format_id,
  sheet_formats = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_format}{formats}
}
\description{
Format function68
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
