% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmt_vecs.R
\name{fmt_fill_patternFill_fgColor_theme}
\alias{fmt_fill_patternFill_fgColor_theme}
\title{Format function20
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.}
\usage{
fmt_fill_patternFill_fgColor_theme(format_id_vec = local_format_id,
  sheet_formats = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_format}{formats}
}
\description{
Format function20
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
