% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmt_vecs.R
\name{fmt_font_underline}
\alias{fmt_font_underline}
\title{Format function8
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.}
\usage{
fmt_font_underline(format_id_vec = local_format_id,
  sheet_formats = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_formats}{formats}
}
\description{
Format function8
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
