% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmt_vecs.R
\name{fmt_protection_locked}
\alias{fmt_protection_locked}
\title{Format function100
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.}
\usage{
fmt_protection_locked(format_id_vec = local_format_id,
  sheet_formats = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_formats}{formats}
}
\description{
Format function100
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
