% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_get_col_groups.R
\name{get_col_groups}
\alias{get_col_groups}
\title{Get column groups}
\usage{
get_col_groups(sheet, value_ref, formats,
  group_col_headers_by = c("bolding"),
  col_header_fill = "local_format_id",
  default_col_header_direction = default_col_header_direction)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}

\item{value_ref}{data frame representing corners of numeric cells in excel sheet}

\item{formats}{format object read in by \code{tidyxl::xlsx_cells}}
}
\description{
This function:
1. Identifies which cells are likely to be headers
2. groups them according to their indenting, bold and italic formatting
3. Specifies the unpivotr function specifying the direction of the header w.r.t. table data
}
