% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{get_corner_cell_refs}
\alias{get_corner_cell_refs}
\title{get_corner_cell_refs}
\usage{
get_corner_cell_refs(sheet)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}
}
\description{
Identifies corners of sheet
}
