% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020_get_formats.R
\name{get_indenting}
\alias{get_indenting}
\title{Get formats
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.}
\usage{
get_indenting(format_id = local_format_id, sheet_format = formats)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Get formats
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
