% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_get_orientation_dfs.R
\name{get_orientation_df}
\alias{get_orientation_df}
\alias{get_orientations_df}
\title{Produces a data frame with a row for each col/row combination in a the provided range
and with direction information extracted from the name provided by the orientation object.}
\usage{
get_orientation_df(orientation)

get_orientations_df(orientations)
}
\arguments{
\item{orientation}{a vector of length one representing a direction (N,NNW,W, WNW, etc.) that has a corresponding name that represents a range (e.g A2:B11).
For example, \link{c("C2:D2" = "N")}}

\item{orientation}{a vector of abritrary length representing  directions (N,NNW,W, WNW, etc.) that have a corresponding names that represent ranges (e.g A2:B11).
For example, \link{c("C1:D1" = "N",  "C2:D2" = "N", "B3:B4" = "W" ,  "A3:A4" = "W",  "c3:d4" = "data")}}
}
\description{
This function produces a data frame with a row for each col/row combination in a the provided range
and with direction information extracted from the name provided by the orientation object.#'

This function produces a data frame with a row for each col/row combination in a the provided range
and with direction information extracted from the name provided by the orientation object.#'
}
\examples{



}
