% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{get_range_df}
\alias{get_range_df}
\title{Produces a data frame with a row for each col/row combination in a the provided range.}
\usage{
get_range_df(range)
}
\arguments{
\item{range}{a string representing a range in standard excel format. For example, "A4:Z15"}
}
\description{
This function produces a data frame with a row for each col/row combination in a the provided range.
}
