% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_get_range_dfs.R
\name{get_range_dfs}
\alias{get_range_dfs}
\title{Produces a data frame with a row for each col/row combination in a the provided range.}
\usage{
get_range_dfs(range)
}
\arguments{
\item{range}{a string representing a range in standard excel format. For example, "A4:B15, C11:G22"}
}
\description{
This function produces a data frame with a row for each col/row combination in a the provided range.
It handles instances where the range is a union of multiple ranges separate by commas.
}
