% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_get_row_groups.R
\name{get_row_groups}
\alias{get_row_groups}
\title{Get Rowname DF}
\usage{
get_row_groups(sheet, value_ref, formats,
  .groupings = groupings(fmt_alignment_indent),
  header_fill = "local_format_id",
  default_row_header_direction = default_row_header_direction,
  table_data = tabledata, filter_headers_by = filter_headers_by_temp,
  min_header_index = min_header_index_temp)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}

\item{value_ref}{data frame representing corners of numeric cells in excel sheet}

\item{formats}{format object read in by \code{tidyxl::xlsx_cells}}

\item{col_groups}{format object read in by \code{tidyxl::xlsx_cells}}

\item{added_row_groups}{format object read in by \code{tidyxl::xlsx_cells}}
}
\description{
This function:
1. Identifies which cells are likely to be row names
2. groups them according to their indenting, bold and italic formatting
3. Specifies the unpivotr function specifying the direction of the header w.r.t. table data
}
