% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_row_groups.R
\name{get_row_sum}
\alias{get_row_sum}
\title{Get row sum}
\usage{
get_row_sum(data, sheet)
}
\arguments{
\item{data}{a row_name_df object}

\item{sheet}{a row_name_df object}
}
\description{
This function is used to identify whether rows have a Wester or NNW orientation to data
}
