% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_references.R
\name{get_value_references}
\alias{get_value_references}
\title{Get table references}
\usage{
get_value_references(sheet, manual_value_references)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}

\item{manual_value_references}{sheet object read in by \code{tidyxl::xlsx_cells}}
}
\description{
Identifies rectangle of numeric cells in table.
}
