% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl_fmts.R
\name{h_alignment}
\alias{h_alignment}
\title{Add formatting information from the h_alignment format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' h_alignment formatting.}
\usage{
h_alignment(format_id_vec = local_format_id, sheet_format = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_format}{formats}
}
\description{
Add formatting information from the h_alignment format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' h_alignment formatting.
}
