% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fmt_vecs.R
\name{indenting}
\alias{indenting}
\title{Format function1
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.}
\usage{
indenting(format_id_vec = local_format_id, sheet_format = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_format}{formats}
}
\description{
Format function1
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
