% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.R
\name{isolate_sentinels}
\alias{isolate_sentinels}
\title{Isolate sentinel values into a separate column}
\usage{
isolate_sentinels(.data, sentinels, col, into = sentinel)
}
\description{
Isolate sentinel values into a separate column
}
\examples{
isolate_sentinels(Formaldehyde, 0.3, carb)
isolate_sentinels(Formaldehyde, 0.3, carb, flag)
}
