% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_header.R
\name{join_header}
\alias{join_header}
\alias{N}
\alias{E}
\alias{S}
\alias{W}
\alias{NNW}
\alias{NNE}
\alias{SSE}
\alias{SSW}
\alias{WNW}
\alias{ENE}
\alias{ESE}
\alias{WSW}
\alias{ABOVE}
\alias{LEFT}
\alias{BELOW}
\alias{RIGHT}
\title{Join data cells to headers}
\usage{
join_header(data_cells, header_cells, direction, corner_cells = NULL,
  drop = TRUE)

N(data_cells, header_cells, drop = TRUE)

E(data_cells, header_cells, drop = TRUE)

S(data_cells, header_cells, drop = TRUE)

W(data_cells, header_cells, drop = TRUE)

NNW(data_cells, header_cells, drop = TRUE)

NNE(data_cells, header_cells, drop = TRUE)

SSE(data_cells, header_cells, drop = TRUE)

SSW(data_cells, header_cells, drop = TRUE)

WNW(data_cells, header_cells, drop = TRUE)

ENE(data_cells, header_cells, drop = TRUE)

ESE(data_cells, header_cells, drop = TRUE)

WSW(data_cells, header_cells, drop = TRUE)

ABOVE(data_cells, header_cells, corner_cells = NULL, drop = TRUE)

LEFT(data_cells, header_cells, corner_cells = NULL, drop = TRUE)

BELOW(data_cells, header_cells, corner_cells = NULL, drop = TRUE)

RIGHT(data_cells, header_cells, corner_cells = NULL, drop = TRUE)
}
\arguments{
\item{data_cells}{Data frame of data cells with at least the columns 'row'
and 'column', which are \code{numeric} or \code{integer}.}

\item{header_cells}{Data frame of header cells with at least the columns
'row' and 'column', which are numeric/integer vectors.}

\item{direction}{The name of a function that joins headers to data cells, one
of \code{N}, \code{E}, \code{S}, \code{W}, \code{NNW}, \code{NNE}, \code{ENE}, \code{ESE}, \code{SSE}, \code{SSW}. \code{WSW},
\code{WNW}, \code{ABOVE}, \code{BELOW}, \code{LEFT} and \code{RIGHT}.  See 'details'.}

\item{corner_cells}{Data frame of cells in one row or one column, giving the
corner of the area that a header applies to.  "BELOW" and "LEFT".  For
example, \code{corner_cells} could be cells with borders on one side.  This is
useful when the nearest header might be the wrong header because it lies on
the other side of a border.}

\item{drop}{Logical vector length 1. Whether data cells that can't be
associated with a header should be dropped.  Default: \code{TRUE}.}
}
\description{
Data cells in a table are associated with header cells by proximity.
\code{\link[=join_header]{join_header()}} and the shorthand functions \code{\link[=NNW]{NNW()}}, \code{\link[=ABOVE]{ABOVE()}}, etc., join a
data frame of data cells to a data frame of header cells, choosing the
nearest header cells in the given direction.
}
\details{
Headers are associated with data by proximity in a given direction.  The
directions are mapped to the points of the compass, where 'N' is north (up),
'E' is east (right), and so on.  \code{\link[=join_header]{join_header()}} finds the nearest header to
a given data cell in a given direction, and joins it to the data cell.

The most common directions to search are 'NNW' (for left-aligned headers at
the top of the table) and 'WNW' for top-aligned headers at the side of the
table.

The full list of available directions is 'N', 'E', 'S', 'W', 'NNW', 'NNE',
'ENE', 'ESE', 'SSE', 'SSW', 'WSW', 'WNW', 'ABOVE', 'BELOW', 'LEFT', 'RIGHT'.
For convenience, these directions are provided as their own functions,
wrapping the concept of 'join_header()'.

The difference between 'N' and 'ABOVE' (and similar pairs of
directions) is that 'N' finds headers directly above the data cell, whereas
'ABOVE' matches the nearest header, whether above-left, above-right or
directly above the data cell.  This is useful for matching headers that are
not aligned to the edge of the data cells that they refer to.  There can be a
tie in the directions 'ABOVE', 'BELOW', 'LEFT' and 'RIGHT' , causing NAs to
be returned in the place of header values.

The full list of available directions is 'N', 'E', 'S', 'W', 'NNW', 'NNE',
'ENE', 'ESE', 'SSE', 'SSW', 'WSW', 'WNW', 'ABOVE', 'BELOW', 'LEFT', 'RIGHT'.
For convenience, these directions are provided as their own functions,
wrapping the concept of 'join_header()'.
}
\section{Functions}{
\itemize{
\item \code{N}: Join nearest header in the 'N' direction.

\item \code{E}: Join nearest header in the 'E' direction.

\item \code{S}: Join nearest header in the 'S' direction.

\item \code{W}: Join nearest header in the 'W' direction.

\item \code{NNW}: Join nearest header in the 'NNW' direction.

\item \code{NNE}: Join nearest header in the 'NNE' direction.

\item \code{SSE}: Join nearest header in the 'SSE' direction.

\item \code{SSW}: Join nearest header in the 'SSW' direction.

\item \code{WNW}: Join nearest header in the 'WNW' direction.

\item \code{ENE}: Join nearest header in the 'ENE' direction.

\item \code{ESE}: Join nearest header in the 'ESE' direction.

\item \code{WSW}: Join nearest header in the 'WSW' direction.

\item \code{ABOVE}: Join nearest header in the 'ABOVE' direction.

\item \code{LEFT}: Join nearest header in the 'ABOVE' direction.

\item \code{BELOW}: Join nearest header in the 'ABOVE' direction.

\item \code{RIGHT}: Join nearest header in the 'ABOVE' direction.
}}

\examples{
library(dplyr)
# Load some pivoted data
(x <- purpose$`NNW WNW`)
# Make a tidy representation
cells <- tidy_table(x)
cells <- cells[!is.na(cells$chr), ]
head(cells)
# Select the cells containing the values
datacells <-
  filter(cells, row >= 3, col >= 3) \%>\%
  transmute(row, col, count = as.integer(chr))
head(datacells)
# Select the headers
qualification <-
  filter(cells, col == 1) \%>\%
  select(row, col, qualification = chr)
age <-
  filter(cells, col == 2) \%>\%
  select(row, col, age = chr)
gender <-
  filter(cells, row == 1) \%>\%
  select(row, col, gender = chr)
satisfaction <-
  filter(cells, row == 2) \%>\%
  select(row, col, satisfaction = chr)
# From each data cell, search for the nearest one of each of the headers
datacells \%>\%
  NNW(gender) \%>\%
  N(satisfaction) \%>\%
  WNW(qualification) \%>\%
  W(age) \%>\%
  select(-row, -col)

# The `drop` argument controls what happens when for some cells there is no
# header in the given direction. When `drop = TRUE` (the default), cells that
# can't be joined to a header are dropped.  Otherwise they are kept.
N(datacells, gender)
N(datacells, gender, drop = FALSE)
}
