% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_data.R
\name{locate_data}
\alias{locate_data}
\title{Locates data cells}
\usage{
locate_data(sheet = NULL, ...)
}
\arguments{
\item{sheet}{a data frame produced by xlsx_cells}

\item{...}{a filter expression that identifies data cells.}
}
\description{
This function identifies which cells is a tidyxl dataframe represent data cells.
It removes data cells from data frame and stores them in an attribute of the resulting tidyxl data frame.
}
\examples{

# Read in tidyxl data frame
 xl_df <-  
  unpivotr_example("worked-examples.xlsx") \%>\% 
 
Visually inspect the result 
xl_df \%>\% plot_cells 
 
}
