% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{locate_data_test}
\alias{locate_data_test}
\title{Data frame used for testing locate_data()}
\format{a dataframe}
\usage{
locate_data_test
}
\description{
Data frame used for testing locate_data()
}
\keyword{datasets}
