% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{locate_groups_test}
\alias{locate_groups_test}
\title{Data frame used for testing locate_groups()}
\format{a dataframe}
\usage{
locate_groups_test
}
\description{
Data frame used for testing locate_groups()
}
\keyword{datasets}
