% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatate_header_groups_if.R
\name{locate_header_groups}
\alias{locate_header_groups}
\title{get tidyABS components}
\usage{
locate_header_groups(sheet = NULL, ..., type = "both",
  .groupings = groupings(fmt_alignment_indent),
  default_col_header_direction = "N",
  default_row_header_direction = "W", header_fill = "local_format_id")
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }

}
