% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate_if}
\alias{locate_if}
\title{Conditionally adds direction annotations to tidyxl data frame}
\usage{
locate_if(
  cells,
  ...,
  direction,
  name,
  values = NULL,
  types = data_type,
  formatters = list(),
  drop_na = TRUE
)
}
\arguments{
\item{cells}{a tidyxl data frame}

\item{...}{expression to filter for headers}

\item{direction}{direction of new headers}

\item{name}{of variable associated with header}

\item{values}{values}

\item{types}{types}

\item{formatters}{formatters}

\item{drop_na}{drop_na}
}
\description{
This function conditionally adds direction annotations to tidyxl data frame.
}
\examples{

# Read in a formatted tidyxl data frame. 
xl_df <- 
unpivotr_example("worked-examples.xlsx") \%>\%
 xlsx_cells_fmt(sheets = "pivot-annotations") 
 
# Identfy which cells are data cells using `locate_data` 
 
 xl_df <- xl_df \%>\% locate_data(data_type == "numeric") 
 
 # Identify header cells, indicating their direction with respect to the data and a name for their column in the final tidy data frame.
 xl_df <- 
  xl_df \%>\% 
   locate(direction = "WNW", name = subject_type) \%>\% 
   locate(direction = "W", name = subject) \%>\%
   locate(direction = "NNW", name = gender) \%>\% 
   locate(direction = "N", name = name)  
   
 # Use `migrate` to reshape the data frame such that each data cells has its own row and each header variable has its own column.  
 xl_df \%>\% migrate()

}
