% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0000_behead_dc.R, R/0000_locate_if.R,
%   R/0000_migrate_ndc.R, R/050_migrate.R
\name{migrate}
\alias{migrate}
\alias{behead_dc}
\alias{locate}
\alias{locate_if}
\alias{migrate_old}
\title{Transform a tidyxl data frame with directions to a tidy data frame that has a column for each header label.}
\usage{
behead_dc(cells, direction, name, values = NULL, types = data_type,
  formatters = list(), drop_na = TRUE)

locate(cells, direction, name, values = NULL, types = data_type,
  formatters = list(), drop_na = TRUE)

locate_if(cells, ..., direction, name, values = NULL,
  types = data_type, formatters = list(), drop_na = TRUE)

migrate(orientated_df, numeric_value = FALSE)

migrate_old(orientated_df, numeric_value = FALSE)
}
\arguments{
\item{orientated_df}{a tidyxl data frame with a \code{.direction} and \code{.header.group} columns}

\item{orientated_df}{a tidyxl data frame with a \code{.direction} and \code{.header.group} columns}

\item{orientated_df}{a tidyxl data frame with a \code{.direction} and \code{.header.group} columns}

\item{orientated_df}{a tidyxl data frame with a \code{.direction} and \code{.header.group} columns}

\item{orientated_df}{a tidyxl data frame with a \code{.direction} and \code{.header.group} columns}
}
\description{
This function is to be used following \code{\link[unpivotr:migrate]{unpivotr::migrate()}}.
It transforms a tidyxl data frame with directions to a tidy data frame that has a column for each header label.

This function is to be used following \code{\link[unpivotr:migrate]{unpivotr::migrate()}}.
It transforms a tidyxl data frame with directions to a tidy data frame that has a column for each header label.

This function is to be used following \code{\link[unpivotr:migrate]{unpivotr::migrate()}}.
It transforms a tidyxl data frame with directions to a tidy data frame that has a column for each header label.

This function is to be used following \code{\link[unpivotr:migrate]{unpivotr::migrate()}}.
It transforms a tidyxl data frame with directions to a tidy data frame that has a column for each header label.

This function is to be used following \code{\link[unpivotr:migrate]{unpivotr::migrate()}}.
It transforms a tidyxl data frame with directions to a tidy data frame that has a column for each header label.
}
