% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{migrate_test}
\alias{migrate_test}
\title{Data frame used for testing migrate()}
\format{a dataframe}
\usage{
migrate_test
}
\description{
Data frame used for testing migrate()
}
\keyword{datasets}
