% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{mutate}
\alias{mutate}
\title{Create or transform variables}
\usage{
mutate(.data, ...)
}
\arguments{
\item{...}{Name-value pairs of expressions, each with length 1 or the same}
}
\description{
This is mutate from dply
}
