% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{mutate_fmt}
\alias{mutate_fmt}
\title{Select keeping formats}
\usage{
mutate_fmt(df, ...)
}
\arguments{
\item{df}{a data frame created by tidyxl::xlsx_cells}

\item{...}{select input.}
}
\description{
A wrapper for dplyr::mutate that retains formatting information
}
\examples{
 
library(tidyverse)

annotated_df <- 
unpivotr_example("newresconst.xlsx") \%>\% 
  xlsx_cells_fmt(sheets = "Table 1 - Permits") \%>\%
  append_fmt(fmt_font_bold) \%>\% 
  filter_fmt(row < min(row[str_detect(character,"RSE")],na.rm = TRUE)) \%>\% 
  mutate_fmt(double_height = height *2) \%>\% 
  locate_data(data_type == "numeric" & col > 1) \%>\%
  locate_groups(direction = "W", 
                .groupings = groupings(is.na(numeric)), 
                .hook_if = hook(any(data_type == "numeric"))) \%>\% 
  locate_groups(direction = "N", header_fill = "style")  

 annotated_df \%>\% plot_cells()
}
