% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_language_expressions}
\alias{name_language_expressions}
\title{Get name from a language object}
\usage{
name_language_expressions(x, prefix)
}
\arguments{
\item{x}{a language}
}
\description{
This is an internal function that adds a prefixed name to a quosure so that variables added to a data frame using this quosure have a predictable name.
}
