% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_string_expressions}
\alias{name_string_expressions}
\title{Get name from a vector of spreadsheet range strings}
\usage{
name_string_expressions(x, prefix)
}
\arguments{
\item{x}{a vector of strings representing spreadsheet ranges}
}
\description{
This is an internal function that adds a prefixed name to a quosure so that variables added to a data frame using this quosure have a predictable name.
}
