% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_symbol_expressions}
\alias{name_symbol_expressions}
\title{Get name from a vector of symbols}
\usage{
name_symbol_expressions(symbols, prefix)
}
\arguments{
\item{x}{a vector of symbols}
}
\description{
This is an internal function that adds a prefixed name to a quosure so that variables added to a data frame using this quosure have a predictable name.
}
