% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_orientate.R
\name{orientate}
\alias{orientate}
\title{Manually add orientation information to a tidyxl data frame.}
\usage{
orientate(tidyxl_df, orientations)
}
\arguments{
\item{tidyxl_df}{a tidyxl data frame}

\item{orientations}{a vector of directions/data indicators with corresponding ranges as names.
For example:  c("C1:D1" = "N",  "C2:D2" = "N", "B3:B4" = "W" ,  "A3:A4" = "W",  "c3:d4" = "data")}
}
\description{
This function adds a two columns: .orientation and .header_group.
.orientation indicates the directional relationship between the current cell and the data in the middle of the table.
.header_group represents the name of the header group.
}
