% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientate_auto.R.R
\name{orientate_auto}
\alias{orientate_auto}
\title{get tidyABS components}
\usage{
orientate_auto(sheet, formats, manual_value_references = NULL,
  added_row_groups = NULL, keep_meta_data = FALSE)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
