% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_directions.R
\name{plot_directions}
\alias{plot_directions}
\title{plot orientations}
\usage{
plot_directions(sheet, interactive = FALSE)
}
\arguments{
\item{orientated_df}{returned by orientated or orientate_df}
}
\description{
This function plots the excel sheet, highlighting the relationship between headers and table values.
}
\examples{

}
