% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040_plot_orientations.R
\name{plot_orientations}
\alias{plot_orientations}
\title{plot orientations}
\usage{
plot_orientations(orientated_df)
}
\arguments{
\item{orientated_df}{returned by orientated or orientate_df}
}
\description{
This function plots the excel sheet, highlighting the relationship between headers and table values.
}
\examples{

}
