% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify.R
\name{rectify}
\alias{rectify}
\alias{print.cell_grid}
\title{Display cells as though in a spreadsheet}
\usage{
rectify(cells, values = NULL, types = data_type)

\method{print}{cell_grid}(x, display = "terminal", ...)
}
\arguments{
\item{cells}{Data frame or tbl, the cells to be displayed.}

\item{values}{Optional. The column of \code{cells} to use as the values of each
cell.  Given as a bare variable name.  If omitted (the default), the \code{types}
argument will be used instead.}

\item{types}{The column of \code{cells} that names, for each cell, which column to
use for the value of the cell.  E.g.  a cell with a character value will have
\code{"character"} in this column.  If this argument is given, then \code{...} will be
ignored.}

\item{x}{The output of \code{\link[=rectify]{rectify()}}}

\item{display}{One of \code{"terminal"} (default), \code{"browser"}, \code{"rstudio"}.  To
display in the browser you must have the \code{DT} package installed.}
}
\description{
Takes the 'melted' output of \code{tidy_table()} or
\code{tidyxl::xlsx_cells()} (each row represents one cell) and projects the cells
into their original positions.  By default this prints to the
terminal/console, but with \code{display = "browser"} or \code{display = "rstudio"} it
will be displayed in the browser or the RStudio viewer pane.

This is for viewing only; the output is not designed to be used in other
functions.

Example: The following cells\preformatted{row col value
  1   1   "a"
  1   2   "b"
  2   1   "c"
  2   2   "d"
}

Would be presented as\preformatted{  1(A) 2(B)
1 "a"  "b"
2 "c"  "d"
}

The letters in the column names are for comparing this view with a
spreadsheet application.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: S3 method for class \code{cell_grid}
}}

\examples{
BOD
x <- tidy_table(BOD, TRUE, TRUE)
x
rectify(x)

# You can choose to use a particular column of the data
rectify(x, chr)
rectify(x, dbl)

# You can also show which row or which column each cell came from, which
# helps with understanding what this function does.
rectify(x, row)
rectify(x, col)

# Empty rows and columns up to the first occupied cell are dropped, but the
# row and column names reflect the original row and column numbers.
x$row <- x$row + 5
x$col <- x$col + 5
rectify(x)
#
# Print in the browser or in the RStudio viewer pane
\dontrun{
  y <- rectify(mtcars_tidy, chr, dbl)
  print(y, "browser")
  print(y, "rstudio")
}
}
