% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0000_helper-functions.R,
%   R/0000_reduce_mutated.R
\name{reduce_mutated}
\alias{reduce_mutated}
\title{get tidyABS components}
\usage{
reduce_mutated(df, form_list, format)

reduce_mutated(df, form_list, format)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}

\item{header_df}{a data frame containing header cells.}

\item{col_header_fill}{the manner is which blank cells will be filled.}

\item{formats}{the formats associated with the workbook containing the header_df cells.}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.

This function ensures that merged cells are unmerged.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
