% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{select}
\alias{select}
\title{Select/rename variables by name}
\usage{
select(.data, ...)
}
\arguments{
\item{...}{One or more unquoted expressions separated by commas.}
}
\description{
This is select from dply.
}
