% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatter.R
\name{spatter}
\alias{spatter}
\title{Spatter a key-values set across multiple columns.}
\usage{
spatter(.data, key, values, types = data_type)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Spatter a key-values set across multiple columns.
}
