% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{string_range_to_filter_vars}
\alias{string_range_to_filter_vars}
\title{get tidyABS components}
\usage{
string_range_to_filter_vars(sheet, table_range)
}
\arguments{
\item{sheet}{sheet nominated for tidying}

\item{table_range}{table_range}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
