% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_to_range}
\alias{string_to_range}
\title{Convert spreadsheet range to a vector of row-col strings.}
\usage{
string_to_range(string)
}
\arguments{
\item{x}{a string representing a spreadsheet range}
}
\description{
This is an internal function that converts a spreadsheet range to a vector of row-col strings.
}
