% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symbol_expressions_to_quosures}
\alias{symbol_expressions_to_quosures}
\title{Convert symbol to a filtering quosure}
\usage{
symbol_expressions_to_quosures(symbol_expression, environ)
}
\arguments{
\item{symbol_expression}{a symbol expression}
}
\description{
Convert an expression that represents a fmt_ function to a quosure
that can be used to create a variable that returns TRUE if row-col
combination is in the range.
}
