% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010_unmerge_cells.R
\name{unmerge_cells}
\alias{unmerge_cells}
\title{Fill in blanks}
\usage{
unmerge_cells(sheet, strict_merging = T)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}
}
\description{
This function ensures that merged cells are unmerged.
}
