% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack}
\alias{unpack}
\title{Unpack cell values from one list-column into separate columns per data type}
\usage{
unpack(.data, values = value, name = data_type)
}
\description{
Unpack cell values from one list-column into separate columns per data type
}
