% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\docType{data}
\name{worked_example_datacells}
\alias{worked_example_datacells}
\title{A data frame used for testing.}
\format{A character vector of length 702}
\usage{
worked_example_datacells
}
\description{
A dataframe with data_cells added correctly
}
\keyword{datasets}
