% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\docType{data}
\name{worked_example_locate_if}
\alias{worked_example_locate_if}
\title{A data frame used for testin locate_if}
\format{A character vector of length 702}
\usage{
worked_example_locate_if
}
\description{
A dataframe with location information added correctly using lacate_if
}
\keyword{datasets}
