% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl_xlsx_cells_fmt.R
\name{xlsx_cells_fmt}
\alias{xlsx_cells_fmt}
\title{Produce dataframe from excel file, storing formats as an attribute.}
\usage{
xlsx_cells_fmt(path, sheets = NA)
}
\arguments{
\item{path}{path to excel file}

\item{sheets}{sheet to be extracted.}
}
\description{
This function produces a data frame from excel file, storing formats as an attribute
}
\examples{
print("todo")
}
