% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{add_variable_if_missing}
\alias{add_variable_if_missing}
\title{An internal function to add a variable to a tidyxl data frame in that variable is missing}
\usage{
add_variable_if_missing(sheet, var)
}
\arguments{
\item{sheet}{sheet nominated for tidying}

\item{var}{name of the var to be added.}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
