% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_variable_if_missing}
\alias{add_variable_if_missing}
\title{Add a variable to a tidyxl data frame if that variable is missing}
\usage{
add_variable_if_missing(sheet, var)
}
\arguments{
\item{sheet}{sheet nominated for tidying}

\item{var}{name of the var to be added as a character string.}
}
\description{
This is an internal function that takes a tidyxl data frame and a variable name to be added if they are missing.
}
