% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{append_fmt}
\alias{append_fmt}
\title{Add multiple format variables}
\usage{
append_fmt(cells, ...)
}
\arguments{
\item{cells}{a data frame created by tidyxl::xlsx_cells}

\item{...}{fmt_* function.}
}
\description{
Add a format variable.
}
\examples{
print("todo")
}
