% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_name_to_quosure}
\alias{append_name_to_quosure}
\title{Give quosure a name}
\usage{
append_name_to_quosure(x, prefix = "grp_")
}
\arguments{
\item{x}{quosure}

\item{prefix}{prefix to be added in name}
}
\description{
This is an internal function that adds a prefixed name to a quosure so that variables added to a data frame using this quosure have a predictable name.
}
