% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behead_groups_if.R
\name{behead_groups_if}
\alias{behead_groups_if}
\title{Behead multiple header groups}
\usage{
behead_groups_if(sheet = NULL, ..., direction = "W",
  .groupings = groupings(fmt_alignment_indent),
  .hook_if = hook(any(FALSE)), .hook_if_rev = hook(any(FALSE)),
  default_col_header_direction = "N",
  default_row_header_direction = "W", header_fill = "local_format_id")
}
\arguments{
\item{sheet}{data frame created by xlsx_cells}

\item{...}{filter expression identifying header cells.}

\item{direction}{a string indicating which type of headers are to be labelled. Options include compass direction or up/down/left/right.}

\item{.groupings}{expressions representing how header cells are differentiated. Most naturally works with fmt_* functions.}

\item{.hook_if}{expression determining whether directions are hooked.}

\item{.hook_if_rev}{expression determining whether directions are reverse hooked.}

\item{default_col_header_direction}{Indicates which direction is given to col headers by default. Only need if "NNW" is required, rather than "N".}

\item{default_row_header_direction}{Indicates which direction is given to row headers by default. Only need if "WNW" is required, rather than "W".}

\item{header_fill}{deals with merged cells. Fills in neighbouring cells if they have the same "local_format_id", "style" or are within "borders".}
}
\description{
Beheads multiple headers defined according to expressions in .groupings.
}
\examples{
print("todo")
}
