% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{direction_plot_interactive}
\alias{direction_plot_interactive}
\title{Matches directions to unicode arrows for an interative chart}
\format{a dataframe}
\usage{
direction_plot_interactive
}
\description{
A data frame with .direction, .arrow and .rate columns.
Because rotation of unicode arrows is not possible with plotly,
a separate set of straight arrows is used for interative use of plot_cells.
}
\keyword{datasets}
