% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{direction_plot_noninteractive}
\alias{direction_plot_noninteractive}
\title{Matches directions to unicode arrows for an interative chart}
\format{a dataframe}
\usage{
direction_plot_noninteractive
}
\description{
A data frame with .direction, .arrow and .rate columns
Since ggplot allows for rotation of plotted text,
this datafrom is comprised of right-angled arrows,
some of which are rotated by plot cells.
}
\keyword{datasets}
