% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{enhead_tabledata}
\alias{enhead_tabledata}
\title{Fill in blanks}
\usage{
enhead_tabledata(header_data, direction, values = tabledata)
}
\arguments{
\item{header_df}{a data frame containing header cells.}

\item{col_header_fill}{the manner is which blank cells will be filled.}

\item{formats}{the formats associated with the workbook containing the header_df cells.}
}
\description{
This function ensures that merged cells are unmerged.
}
