% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{enhead_tabledata}
\alias{enhead_tabledata}
\title{Fill in blanks}
\usage{
enhead_tabledata(header_data, direction, values = tabledata)
}
\arguments{
\item{header_data}{a data frame containing header cells.}

\item{direction}{direction suplied to enhead.}

\item{values}{the data values.}
}
\description{
This function ensures that merged cells are unmerged.
}
