% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_blanks_in_col_headers}
\alias{fill_blanks_in_col_headers}
\title{Unmerge and fill in blanks for col headers.}
\usage{
fill_blanks_in_col_headers(header_df, header_fill = c("style",
  "local_format_id", "borders"), formats)
}
\arguments{
\item{header_df}{a data frame containing header cells.}

\item{header_fill}{the criteria according to which blank cells will be filled.}

\item{formats}{the formats of the workbook associated with the tidyxl data frame.}
}
\description{
This is an internal function that function ensures that merged cells are unmerged.
}
