% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_blanks_in_headers}
\alias{fill_blanks_in_headers}
\title{Unmerge and fill in blanks for headers}
\usage{
fill_blanks_in_headers(header_df, header_fill = c("style",
  "local_format_id", "borders"), formats, direction)
}
\arguments{
\item{header_df}{a data frame containing header cells.}

\item{header_fill}{he criteria according to which blank cells will be filled.}

\item{formats}{the formats of the workbook associated with the tidyxl data frame.}

\item{direction}{the direction in which headers are filled.}
}
\description{
This is an internal function that function ensures that merged cells are unmerged.
}
