% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl_fmts.R
\name{fmt_font_bold}
\alias{fmt_font_bold}
\title{Add formatting information from the fmt_font_bold format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' font_bold formatting.}
\usage{
fmt_font_bold(format_id_vec = local_format_id, sheet_formats = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_formats}{formats}
}
\description{
Add formatting information from the fmt_font_bold format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' font_bold formatting.
}
