% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_corner_cell_refs}
\alias{get_corner_cell_refs}
\title{Get the corner references of a tidyxl data frame.}
\usage{
get_corner_cell_refs(sheet)
}
\arguments{
\item{sheet}{a tidyxl data frame.}
}
\description{
This is an internal function that takes a tidyxl data frame a returns a dataframe indicating the maximum and minimum corners.
}
