% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{get_header_index}
\alias{get_header_index}
\title{Get current column number}
\usage{
get_header_index(labels, regex_term = "^col_header")
}
\arguments{
\item{labels}{list of labels.}

\item{regex_term}{prefix for col_header.}
}
\description{
This function gets the maximum col_group number
}
