% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_df.R
\name{get_meta_groups}
\alias{get_meta_groups}
\title{get metadata df}
\usage{
get_meta_groups(sheet, value_ref, formats,
  .groupings = groupings(fmt_alignment_indent),
  header_fill = "local_format_id",
  default_row_header_direction = default_row_header_direction,
  table_data = tabledata, filter_headers_by = filter_headers_by_temp,
  min_header_index = min_header_index_temp)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}

\item{value_ref}{data frame representing corners of numeric cells in excel sheet}

\item{formats}{format object read in by \code{tidyxl::xlsx_cells}}

\item{col_groups}{format object read in by \code{tidyxl::xlsx_cells}}
}
\description{
This function:
1. Identifies which cells are likely to to contains meta data (in top right corner)
2. groups them according to their indenting, bold and italic formatting
3. Specifies the unpivotr function specifying the direction of the header w.r.t. table data
}
