% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatate_header_groups.R
\name{get_range_dfs}
\alias{get_range_dfs}
\alias{hook}
\title{Supply an expression to identify which header groups are hooked.}
\usage{
hook(...)
}
\arguments{
\item{...}{expression applied to a  identifies which header groups are hooked}
}
\description{
This function is used with the \code{.hook_if} or \code{.hook_if_rev} arguments in the \code{locate_groups} function.
This expression must evaluate to a single value for each header group.
For example, see below that any(fmt_alignment_indent == 0) is used rather that fmt_alignment_indent == 0.

It passes an expression to \code{dplyr::summarise} that identifies which header groups are hooked - for example swiched from N to NNW.
See the the \code{locate_groups} documentation for more information and an example.
}
\examples{

# Read in tidyxl data frame
xl_df <-  
unpivotr_example("worked-examples.xlsx") \%>\%
xlsx_cells_fmt(sheets = "pivot-hierarchy") \%>\%
append_fmt(fmt_alignment_indent) 

# Add location annotations
xl_df <- 
 xl_df \%>\% 
  locate_data(data_type == "numeric") \%>\%
  locate_groups(direction = "W",
                .groupings = groupings(fmt_alignment_indent),
                .hook_if =     hook(any(fmt_alignment_indent == 0))) \%>\%
  locate(direction = "N", name = student) 
  
# Use `migrate` to reshape the data frame such that each data cells has its own row and each header variable has its own column.  
 xl_df \%>\% migrate()
}
