% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatate_header_groups_if.R
\name{get_range_dfs}
\alias{get_range_dfs}
\alias{hook}
\title{Supply an expression to identify which header groups are hooked.}
\usage{
hook(...)
}
\arguments{
\item{...}{expression applied to a  identifies which header groups are hooked}
}
\description{
This function is used with the \code{.hook_if} or \code{.hook_if_rev} arguments in the \code{locate_groups} function.
It passes an expression to \code{dplyr::summarise} that identifies which header groups are hooked - for example swiched from N to NNW.
See the the \code{locate_groups} documentation for more information and an example.
}
\examples{
print("todo")
}
