% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouings.R
\name{groupings}
\alias{groupings}
\title{Groupings}
\usage{
groupings(...)
}
\arguments{
\item{dots}{gouring expressions}
}
\description{
This functions passes grouping expressions to the .groupings argument of locate_groups. It works most naturually with fmt_* functions.
}
