% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatate_header_groups_if.R
\name{groupings}
\alias{groupings}
\title{Groupings}
\usage{
groupings(...)
}
\arguments{
\item{...}{mutate expression to group headers.}
}
\description{
This functions passes grouping expressions to the .groupings argument of locate_groups. It works most naturally with fmt_* functions.
}
\examples{
print("todo")
}
