% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyxl_fmts.R
\name{indenting}
\alias{indenting}
\title{Add formatting information from the indenting format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' indenting formatting.}
\usage{
indenting(format_id_vec = local_format_id, sheet_format = formats)
}
\arguments{
\item{format_id_vec}{local format id vector}

\item{sheet_format}{formats}
}
\description{
Add formatting information from the indenting format object
This function uses the format object created by \code{xlsx_formats} along with `local_format_id`` to create a vector representing cells' indenting formatting.
}
