% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.R
\name{isolate_sentinels}
\alias{isolate_sentinels}
\title{Move sentinel values into a separate column leaving NA behind}
\usage{
isolate_sentinels(.data, sentinels, col, into = "sentinel")
}
\description{
Move sentinel values into a separate column leaving NA behind
}
\examples{
isolate_sentinels(Formaldehyde, 0.3, carb)
isolate_sentinels(Formaldehyde, 0.3, carb, flag)
}
