% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Add direction annations to tidyxl data frame}
\usage{
locate(cells, direction, name, values = NULL, types = data_type,
  formatters = list(), drop_na = TRUE)
}
\arguments{
\item{cells}{a tidyxl data frame}

\item{direction}{direction of new headers}

\item{name}{of variable associated with header}

\item{values}{values}

\item{types}{types}

\item{formatters}{formatters}

\item{drop_na}{drop_na}
}
\description{
This adds direction annations to tidyxl data frame.
}
\examples{
print("todo")
}
