% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_locate_data.R
\name{locate_data}
\alias{locate_data}
\title{get tidyABS components}
\usage{
locate_data(sheet = NULL, ... = !is.na(numeric))
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}

\item{filter}{condition idnetifying cells.}
}
\description{
Moves data cells to an attribute of the tidyxl data frame.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
