% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_data.R
\name{locate_data}
\alias{locate_data}
\title{Locates data cells}
\usage{
locate_data(sheet = NULL, ...)
}
\arguments{
\item{sheet}{data frame produced by xlsx_cells}

\item{...}{filter expression that identifies data cells.}
}
\description{
Removes data cells from data frame and stores in an attribute.
}
\examples{
print("todo")
}
