% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate_if}
\alias{locate_if}
\title{Conditionally adds direction annotations to tidyxl data frame}
\usage{
locate_if(cells, ..., direction, name, values = NULL,
  types = data_type, formatters = list(), drop_na = TRUE)
}
\arguments{
\item{cells}{a tidyxl data frame}

\item{...}{expression to filter for headers}

\item{direction}{direction of new headers}

\item{name}{of variable associated with header}

\item{values}{values}

\item{types}{types}

\item{formatters}{formatters}

\item{drop_na}{drop_na}
}
\description{
This function conditionally adds direction annotations to tidyxl data frame.
}
\examples{
print("todo")
}
