% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050_migrate.R
\name{migrate}
\alias{migrate}
\title{Transform a tidyxl data frame with orientations to a tidy data frame that has a column for each header label.}
\usage{
migrate(orientated_df, numeric_value = FALSE)
}
\arguments{
\item{orientated_df}{a tidyxl data frame with a \code{.orientation} and \code{.header.group} columns}
}
\description{
This function is to be used following \code{\link[unpivotr:migrate]{unpivotr::migrate()}}.
It transforms a tidyxl data frame with orientations to a tidy data frame that has a column for each header label.
}
