% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Reshape a tidyxl data frame such that each header group has its own column and each data cells has its own row.}
\usage{
migrate(located_df)
}
\arguments{
\item{located_df}{a tidyxl data frame with .direction, .value, .header_label columns.}
}
\description{
This function is used on a tidy data frame that has been annotated with \code{locate_data}, \code{locate_header} or  \code{locate_header_groups}. It creates a column for each header group.
}
\examples{
print("todo")
}
