% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_orientate_auto.R
\name{orientate_auto}
\alias{orientate_auto}
\title{get tidyABS components}
\usage{
orientate_auto(sheet = NULL, formats = NULL, table_range = NULL,
  manual_value_references = NULL, added_row_groups = NULL,
  group_col_headers_by = c("bolding", "italics"),
  group_row_headers_by = c("bolding", "italics", "indenting"),
  col_header_fill = "local_format_id", row_header_fill = "none",
  default_col_header_direction = "N",
  default_row_header_direction = "W", keep_meta_data = FALSE)
}
\arguments{
\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
