% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste3}
\alias{paste3}
\title{paste3}
\usage{
paste3(..., sep = ", ")
}
\arguments{
\item{...}{objects to be pasted together.}

\item{sep}{separating string.}
}
\description{
An internal fucntion that removes NAs from paste.
Source: https://stackoverflow.com/users/1855677/42.
}
