% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells.R
\name{plot_cells}
\alias{plot_cells}
\title{Plots tidyxl data frame in a grid layout}
\usage{
plot_cells(sheet, text = values, interactive = FALSE)
}
\arguments{
\item{sheet}{tidyxl data frame}

\item{text}{text variable to be represented in values plot.}

\item{interactive}{TRUE produces an interactive plotly visualisation.}
}
\description{
This function plots the excel sheet, highlighting the relationship between headers and table values.
}
\examples{

library(tidyverse)

# Create annotated tidyxl data frame
xl_df <- 
unpivotr_example("worked-examples.xlsx") \%>\%
xlsx_cells_fmt(sheets = "pivot-hierarchy") \%>\%
  append_fmt(fmt_alignment_indent) \%>\% 
  locate_data(data_type == "numeric") \%>\%
  locate_groups(direction = "W",
                .groupings = groupings(fmt_alignment_indent),
                .hook_if =     hook(any(fmt_alignment_indent == 0))) \%>\%
  locate(direction = "N", name = student) 
 
# Plot direction annotations  
 xl_df \%>\% plot_cells()
}
