% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_directions.R
\name{plot_cells}
\alias{plot_cells}
\title{Plots tidyxl data frame in a grid layout}
\usage{
plot_cells(sheet, text = values, interactive = FALSE)
}
\arguments{
\item{sheet}{tidyxl data frame}

\item{text}{text variable to be represented in values plot.}

\item{interactive}{TRUE produces an interactive plotly visualisation.}
}
\description{
This function plots the excel sheet, highlighting the relationship between headers and table values.
}
\examples{
print("todo")
}
