% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{reduce_mutated}
\alias{reduce_mutated}
\title{Fill in blanks}
\usage{
reduce_mutated(df, form_list)

reduce_mutated(df, form_list)
}
\arguments{
\item{header_df}{a data frame containing header cells.}

\item{col_header_fill}{the manner is which blank cells will be filled.}

\item{formats}{the formats associated with the workbook containing the header_df cells.}

\item{path}{path to .xlsx file}

\item{sheets}{sheet nominated for tidying}
}
\description{
This function ensures that merged cells are unmerged.

Produces the  tidyABS table components, which store information on column groups, row groups and tabledata.
}
\examples{

 \donttest{tidyABS_example("australian-industry.xlsx") \%>\% process_sheet(sheets = "Table_1")  }



}
