% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_utils.R
\name{reduce_mutated}
\alias{reduce_mutated}
\title{Fill in blanks}
\usage{
reduce_mutated(df, form_list, format)
}
\arguments{
\item{df}{a data frame containing header cells.}

\item{form_list}{list of fomulae}

\item{format}{the formats associated with the workbook containing the header_df cells.}
}
\description{
This function ensures that merged cells are unmerged.
}
