% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reduce_mutated}
\alias{reduce_mutated}
\title{Add a set of variables based on a list of formulae}
\usage{
reduce_mutated(df, form_list, format)
}
\arguments{
\item{df}{a data frame containing header cells.}

\item{form_list}{list of formulae}

\item{format}{the formats associated with the workbook containing the header_df cells.}
}
\description{
This is an internal function that takes a data frame and a list of formulae and appends an arbitrary number of columns to the data frame based on the formulae.
}
