% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{select_fmt}
\alias{select_fmt}
\title{Select keeping formats}
\usage{
select_fmt(df, ...)
}
\arguments{
\item{df}{a data frame created by tidyxl::xlsx_cells}

\item{...}{select input.}
}
\description{
A wrapper for dplyr::select that retains formatting information
}
\examples{
print("todo")
}
