% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_range_to_filter_vars}
\alias{string_range_to_filter_vars}
\title{Create a filtering variable based on a excel string range.}
\usage{
string_range_to_filter_vars(sheet, table_range)
}
\arguments{
\item{sheet}{a tidyxl data frame.}

\item{table_range}{a table_range given by string. For example, "A1:B2"}
}
\description{
This is an internal function that create a variabled prefixed "flt" based on whether cells are within a range specified by an excel string range.
}
