% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmerge_cells.R
\name{unmerge_cells}
\alias{unmerge_cells}
\title{Fill in blanks}
\usage{
unmerge_cells(sheet, strict_merging = T, merge_var)
}
\arguments{
\item{sheet}{sheet object read in by \code{tidyxl::xlsx_cells}}

\item{strict_merging}{only merge on local_format_id.}

\item{merge_var}{the variable determining whether cells a merged.}
}
\description{
This function ensures that merged cells are unmerged.
}
