% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpivotr_example}
\alias{unpivotr_example}
\title{Get path to example data
Unpivotr comes bundled with some example files in its \code{inst/extdata}
directory. This function makes them easy to access.}
\usage{
unpivotr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
Get path to example data
Unpivotr comes bundled with some example files in its \code{inst/extdata}
directory. This function makes them easy to access.
}
